requirements.txt
python-telegram-bot==20.7
asyncpg==0.29.0
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.6.1
aioredis==2.0.1
pytest-asyncio==0.23.5
pytest==8.0.0 